[KAN-118] refactor: renderAvatars 함수를 훅으로 분리
