+++ src/app/components/InformationCard/InformationCard.tsx
@@ -8,13 +8,13 @@ import {
   IconSaveActive,
   IconSaveInactive,
 } from '@/public/icons';
-import Avatar from './Avatar';
 import InfoChip from '../Chip/InfoChip';
 import ProgressBar from '../ProgressBar/ProgressBar';
 import { GatheringParticipantsType } from '@/types/data.type';
 import { MIN_PARTICIPANTS } from '@/constants/common';
 import { formatDate, formatTimeColon } from '@/utils/formatDate';
 import { useSavedGatheringList } from '@/context/SavedGatheringContext';
+import useAvatars from '@/hooks/useAvatars';
 
 interface InformationCardProps {
   title: string;
@@ -47,7 +47,6 @@ const InformationCard = ({
   const [isSaved, setIsSaved] = useState<boolean>(() =>
     checkGatheringSaved(gatheringId, savedGatherings),
   );
-  const [isHovered, setIsHovered] = useState<boolean>(false);
 
   const handleToggleSave = () => {
     setIsSaved((prev) => !prev);
@@ -56,64 +55,7 @@ const InformationCard = ({
     }
   };
 
-  const handleMouseEnter = () => {
-    setIsHovered(true);
-  };
-
-  const handleMouseLeave = () => {
-    setIsHovered(false);
-  };
-
-  // function of setting Avatars with remaining
-  const renderAvatars = () => {
-    const maxVisible = 4;
-    const visibleAvatars = participants
-      .slice(0, maxVisible)
-      .map(({ User }) => (
-        <Avatar
-          key={User.id}
-          id={User.id}
-          name={User.name}
-          image={User.image}
-          className='h-28 w-28'
-        />
-      ));
-
-    if (participantCount > maxVisible) {
-      visibleAvatars.push(
-        <div
-          key='remaining'
-          className='group relative'
-          onMouseEnter={handleMouseEnter}
-          onMouseLeave={handleMouseLeave}
-        >
-          <div className='z-base flex h-28 w-28 items-center justify-center rounded-full bg-gray-200 text-14 font-semibold'>
-            +{participantCount - maxVisible}
-          </div>
-
-          <div
-            className={`absolute left-0 top-full ml-12 mt-2 flex w-max -space-x-6 transition-opacity duration-300 ${
-              isHovered
-                ? 'translate-y-0 opacity-100'
-                : 'translate-y-4 opacity-0'
-            }`}
-          >
-            {participants.slice(maxVisible).map(({ User }) => (
-              <Avatar
-                key={User.id}
-                id={User.id}
-                name={User.name}
-                image={User.image}
-                className={`h-28 w-28 transition-transform duration-300 ${isHovered ? 'translate-y-0' : 'translate-y-4'}`}
-              />
-            ))}
-          </div>
-        </div>,
-      );
-    }
-
-    return visibleAvatars;
-  };
+  const { renderAvatars } = useAvatars({ participants, participantCount });
 
   return (
     <div className='h-full w-full divide-y divide-dashed rounded-[24px] border p-6 text-var-gray-900'>
